ueventd: Fix bootdevice by-name/by-num link creation
Create bootdevice by-name/by-num links only if the block
device is the bootdevice.  This fixes the issue of
bootdevice path being created incorrectly if the first
reported device with by-name/by-num is not the bootdevice.

It also fixes devices in which the bootdevice property
doesn't exist, or the symlink is created by a later step in
the init sequence (such as init.rc itself)

Change-Id: Ifb68080e3149917f9ee88a4a421007f29e8d83d2
Signed-off-by: spezi77 <spezi77@gmx.com>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>