Check against LOG_ID_MAX instead of LOG_ID_KERNEL
Logd currently checks against LOG_ID_KERNEL to determine if a given
log_id is available. While this matches with the comments in log.h to
always keep LOG_ID_KERNEL at the end it does not match with other
checks that Logd makes causing inconsistent behavior. This
inconsistency is noticable on devices with log buffers with id's
greater than LOG_ID_KERNEL and liblog reports that logd is unavailable
while logd would accept the request.

Test: compiles. no functional change

Change-Id: I0302d9e3f2657706fdf778f1e78e5604abf566e2
Signed-off-by: Eliminater74 <eliminater74@gmail.com>